+++ frontend/src/components/common/NetworkStatus.jsx
