+++ frontend/src/components/user/UserDashboard.jsx
@@ -23,8 +23,8 @@ import {
   Eye
 } from 'lucide-react';
 import ReactECharts from 'echarts-for-react';
-import { getElections, getResults, getElectionById, hasVoted } from '../api';
-import { mapUsersToProvinces } from '../utils/provinceUtils';
+import { getElections, getResults, getElectionById, hasVoted } from '../../api';
+import { mapUsersToProvinces } from '../../utils/provinceUtils';
 
 const Dashboard = ({ user }) => {
   const [stats, setStats] = useState({
